requirements.txt
PyYAML~=6.0.1
torch~=2.1.0
numpy
lapack-lite
tempita
dragon4
libdivide
Meson
meson-python
spin
OpenBLAS
LAPACK
GCC
libquadmath~=1.26.1
openai~=0.28.1
rich~=13.6.0
tabulate~=0.9.0
matplotlib~=3.8.0
PyMuPDF~=1.23.7
Pillow~=10.1.0
tqdm~=4.66.1
keyboard~=0.13.5
Kivy~=2.2.1
screeninfo~=0.8.1
PyAutoGUI~=0.9.54
scipy
OpenBLAS
LAPACK
GCC
libquadmath
OpenBLAS
LAPACK
GCC
libquadmath~=1.11.4
natsort~=8.4.0